modularize the project using MVC pattern
