+++ Modules/Python/Flask and MySQL/Practice/Burgers/flask_app/templates/details_page.html
