+++ Modules/Python/Flask and MySQL/Practice/Burgers/server.py
@@ -1,64 +1,5 @@
-from flask import Flask,render_template,redirect,request
-
-from burger import Burger
-app = Flask(__name__)
-
-@app.route('/')
-def index():
-    return render_template("index.html")
-
-@app.route('/create',methods=['POST'])
-def create():
-    data = {
-        "name":request.form['name'],
-        "bun": request.form['bun'],
-        "meat": request.form['meat'],
-        "calories": request.form['calories']
-    }
-    Burger.save(data)
-    return redirect('/burgers')
-
-
-
-
-@app.route('/burgers')
-def burgers():
-    return render_template("results.html",all_burgers=Burger.get_all())
-
-
-@app.route('/show/<int:burger_id>')
-def detail_page(burger_id):
-    data = {
-        'id': burger_id
-    }
-    return render_template("details_page.html",burger=Burger.get_one(data))
-
-@app.route('/edit_page/<int:burger_id>')
-def edit_page(burger_id):
-    data = {
-        'id': burger_id
-    }
-    return render_template("edit_page.html", burger = Burger.get_one(data))
-
-@app.route('/update/<int:burger_id>', methods=['POST'])
-def update(burger_id):
-    data = {
-        'id': burger_id,
-        "name":request.form['name'],
-        "bun": request.form['bun'],
-        "meat": request.form['meat'],
-        "calories": request.form['calories']
-    }
-    Burger.update(data)
-    return redirect(f"/show/{burger_id}")
-
-@app.route('/delete/<int:burger_id>')
-def delete(burger_id):
-    data = {
-        'id': burger_id,
-    }
-    Burger.destroy(data)
-    return redirect('/burgers')
+from flask_app import app
+from flask_app.controllers import burgers
 
 if __name__=="__main__":
     app.run(debug=True)
